RG-2542 Consolidate ReorderHandle and useReorder() -- latter had the only usage
